feat: make hourly data persistent across restarts
During the hourly update, the precipitation and evapotranspiration lists
are saved in a file named `smart_irrigation_<name>.json` in the /config
directory.

Upon startup, the saved data is loaded into the coordinator instance,
but only if auto refresh is enabled (since when auto refresh is enabled,
it's possible to tell if the saved data is stale).

While loading the saved data, if the data is determined to be stale, (ie,
between when the data was last saved and the time that HA is starting
there should've been an auto refresh) the data is discarded.