Refactored code and added set_zone service.
The new service allows to set most of the zone configuration values in
a single zone call.
This change allows to set bucket, duration, throughput, multiplier and
state of the zone.
It defines the handle_set_zone function which replaces previous
handle_set_bucket and handle_set_multiplier methods.
This refactor simplifies code and generalizes the setting of
configuration values.